fix: simplify Webpack config so bundlephobia excludes readable-stream
