+++ src/utils/http.js
@@ -1,4 +1,4 @@
-import { fromNodeStream, fromStream } from './AsyncIterator.js'
+import { fromNodeStream } from './AsyncIterator.js'
 import { asyncIteratorToStream } from './asyncIteratorToStream.js'
 import { collect } from './collect.js'
 
@@ -11,32 +11,6 @@ export async function http ({
   headers = {},
   body
 }) {
-  return global.fetch
-    ? httpBrowser({ core, emitter, emitterPrefix, url, method, headers, body })
-    : httpNode({ core, emitter, emitterPrefix, url, method, headers, body })
-}
-
-async function httpBrowser ({ url, method = 'GET', headers = {}, body }) {
-  // streaming uploads aren't possible yet in the browser
-  if (body) {
-    body = await collect(body)
-  }
-  let res = await global.fetch(url, { method, headers, body })
-  let iter =
-    res.body && res.body.getReader
-      ? fromStream(res.body)
-      : [new Uint8Array(await res.arrayBuffer())]
-  return {
-    url: res.url,
-    method: res.method,
-    statusCode: res.status,
-    statusMessage: res.statusText,
-    body: iter,
-    headers: res.headers
-  }
-}
-
-async function httpNode ({ url, method = 'GET', headers = {}, body }) {
   // If we can, we should send it as a single buffer so it sets a Content-Length header.
   if (body && Array.isArray(body)) {
     body = await collect(body)
